requirements.txt
catalogue==2.0.10
click==8.1.8
colorama==0.4.6
dateparser==1.2.1
et_xmlfile==2.0.0
Faker==37.1.0
joblib==1.5.0
lxml==5.4.0
nltk==3.9.1
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
phonenumbers==9.0.4
pillow==11.2.1
python-dateutil==2.9.0.post0
python-pptx==1.0.2
python-stdnum==1.20
pytz==2025.2
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.2
scrubadub==2.0.1
six==1.17.0
textblob==0.15.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
XlsxWriter==3.2.3
